pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentwonder"
version = "1.0.0"
description = "AgentWonder — YAML-driven agentic workflow platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "httpx>=0.27",
    "google-adk>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
include = ["agentwonder*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
